Some big changes on sizes on buffers, changed a lot of stuff to be hardcoded. Reflections on spheres works again, yay (8 != 8.0). However, triangle rendering still eludes me. I might just know what to do though. Use the padding in triangle desc. Might be something there. Not sure though. Good riddles this is painful. Well, off I go again. Bai

+++ effect/LightHelper.fx
@@ -8,8 +8,8 @@ struct Material
 	float shininess;
 	float reflective;
 	float refractive;
-	int isReflective;
-	int isRefractive;
+	float isReflective;
+	float isRefractive;
 };
 
 struct PointLightData
